+++ python/ray/experimental/__init__.py
@@ -8,7 +8,7 @@ from .features import (
     flush_finished_tasks_unsafe, flush_evicted_objects_unsafe,
     _flush_finished_tasks_unsafe_shard, _flush_evicted_objects_unsafe_shard)
 from .named_actors import get_actor, register_actor
-from .worker import get, wait
+from .api import get, wait
 
 __all__ = [
     "TensorFlowVariables", "flush_redis_unsafe",
